+++ weed/messaging/broker_grpc_server_subscribe.go
@@ -1,0 +1,9 @@
+package messaging
+
+import (
+	"github.com/chrislusf/seaweedfs/weed/pb/messaging_pb"
+)
+
+func (broker *MessageBroker) Subscribe(server messaging_pb.SeaweedMessaging_SubscribeServer) error {
+	panic("implement me")
+}
